+++ src/Makefile.am
@@ -527,10 +527,10 @@ util-validate.h util-affinity.h util-affinity.c \
 util-var.c util-var.h \
 util-var-name.c util-var-name.h \
 util-vector.h \
-win32-syscall.c win32-syscall.h \
-win32-misc.c win32-misc.h \
-win32-service.c win32-service.h \
-win32-syslog.h
+win32/win32-syscall.c win32/win32-syscall.h \
+win32/win32-misc.c win32/win32-misc.h \
+win32/win32-service.c win32/win32-service.h \
+win32/win32-syslog.h
 
 EXTRA_DIST = tests
