+++ src/suricata-common.h
@@ -131,13 +131,13 @@
 #include <syslog.h>
 #else
 #ifdef OS_WIN32
-#include "win32-syslog.h"
+#include "win32/win32-syslog.h"
 #endif /* OS_WIN32 */
 #endif /* HAVE_SYSLOG_H */
 
 #ifdef OS_WIN32
-#include "win32-misc.h"
-#include "win32-service.h"
+#include "win32/win32-misc.h"
+#include "win32/win32-service.h"
 #endif /* OS_WIN32 */
 
 #if HAVE_SYS_TIME_H
