+++ src/util-ioctl.h
@@ -25,7 +25,7 @@
 #include "util-device.h"
 
 #ifdef OS_WIN32
-#include "win32-syscall.h"
+#include "win32/win32-syscall.h"
 #endif
 
 int GetIfaceMTU(const char *pcap_dev);
