+++ src/win32/win32-misc.h
