+++ src/win32/win32-service.c
