+++ src/win32/win32-syslog.h
